+++ lib/pm/adc_esp32.c
@@ -40,49 +40,45 @@ static const adc_bits_width_t width = ADC_WIDTH_BIT_12;
 
 static const adc_atten_t atten = ADC_ATTEN_DB_0;   //11dB attenuation (ADC_ATTEN_DB_11) gives full-scale voltage 3.9V
 static const adc_unit_t unit = ADC_UNIT_1;
-#define DEFAULT_VREF		1100		//Use adc2_vref_to_gpio() to obtain a better estimate
-#define NO_OF_SAMPLES		30			//Multisampling
+#define DEFAULT_VREF        1100        //Use adc2_vref_to_gpio() to obtain a better estimate
+#define NO_OF_SAMPLES        30            //Multisampling
 
-static void checkEfuse(void)
-{
+static void checkEfuse(void) {
     if (esp_adc_cal_check_efuse(ESP_ADC_CAL_VAL_EFUSE_TP) == ESP_OK) {
-        ESP_LOGI(TAG,"eFuse Two Point: Supported\n");
+        ESP_LOGI(TAG, "eFuse Two Point: Supported\n");
     } else {
-        ESP_LOGI(TAG,"Cannot retrieve eFuse Two Point calibration values. Default calibration values will be used.\n");
+        ESP_LOGI(TAG, "Cannot retrieve eFuse Two Point calibration values. Default calibration values will be used.\n");
     }
 }
 
-static void print_char_val_type(esp_adc_cal_value_t val_type)
-{
+static void print_char_val_type(esp_adc_cal_value_t val_type) {
     if (val_type == ESP_ADC_CAL_VAL_EFUSE_TP) {
-        ESP_LOGI(TAG,"Characterized using Two Point Value\n");
+        ESP_LOGI(TAG, "Characterized using Two Point Value\n");
     } else if (val_type == ESP_ADC_CAL_VAL_EFUSE_VREF) {
-        ESP_LOGI(TAG,"Characterized using eFuse Vref\n");
+        ESP_LOGI(TAG, "Characterized using eFuse Vref\n");
     } else {
-        ESP_LOGI(TAG,"Characterized using Default Vref\n");
+        ESP_LOGI(TAG, "Characterized using Default Vref\n");
     }
 }
 
-float analogReadVoltage(uint32_t pin)
-{
+float analogReadVoltage(uint32_t pin) {
     uint32_t adc_reading = 0;
     for (int i = 0; i < NO_OF_SAMPLES; i++) {
         if (unit == ADC_UNIT_1) {
-            adc_reading += adc1_get_raw((adc1_channel_t)channel);
+            adc_reading += adc1_get_raw((adc1_channel_t) channel);
         } else {
             int raw;
-            adc2_get_raw((adc2_channel_t)channel, width, &raw);
+            adc2_get_raw((adc2_channel_t) channel, width, &raw);
             adc_reading += raw;
         }
     }
     adc_reading /= NO_OF_SAMPLES;
     //Convert adc_reading to voltage in mV
     uint32_t voltage = esp_adc_cal_raw_to_voltage(adc_reading, adc_chars);
-    return voltage / 1000.0;
+    return (float) voltage / 1000.0f;
 }
 
-void adcInit(void)
-{
+void adcInit(void) {
     if (isInit) {
         return;
     }
@@ -93,7 +89,7 @@ void adcInit(void)
         adc1_config_width(width);
         adc1_config_channel_atten(channel, atten);
     } else {
-        adc2_config_channel_atten((adc2_channel_t)channel, atten);
+        adc2_config_channel_atten((adc2_channel_t) channel, atten);
     }
 
     //Characterize ADC
@@ -104,7 +100,6 @@ void adcInit(void)
     isInit = true;
 }
 
-bool adcTest(void)
-{
+bool adcTest(void) {
     return isInit;
 }
